+++ app/profile/[userId]/_components/EditProfileModal/DeleteImageButton.tsx
