+++ app/profile/[userId]/_components/MyPageContent.tsx
@@ -1,18 +1,29 @@
 import { MouseEvent, useState } from "react";
 import { useQuery } from "react-query";
+import { useParams } from "next/navigation";
 import { projectListAPI } from "@/app/_apis/projectListAPI";
+import { profileAPI } from "@/app/_apis/ProfileAPI";
 import MypageProjectSection, { MyPageProjectListType } from "./MypageProjectSection";
 import MyPageCategory from "./MyPageCategory";
 import Profile from "./Profile";
 
 function MyPageContent() {
   const [selectCategory, setSelectCategory] = useState<MyPageProjectListType>("myProject");
+  const { userId } = useParams();
+  const { data: currentUserId } = useQuery({
+    queryKey: [`profile-${userId}`],
+    queryFn: async () => {
+      return await profileAPI.getCurrentUserId();
+    },
+  });
   const { data } = useQuery({
     queryKey: [`projectList-${selectCategory}`],
     queryFn: async () => {
       return await projectListAPI.getMyProjectList({ page: 1, size: 8 }, selectCategory);
     },
   });
+
+  console.log("currentUserId", currentUserId);
 
   const handleSelectCategory = (event: MouseEvent<HTMLButtonElement>) => {
     setSelectCategory(event.currentTarget.id as MyPageProjectListType);
