+++ app/profile/[userId]/layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { PropsWithChildren } from "react";
-import Header from "../_components/Header/Header";
+import Header from "../../_components/Header/Header";
 
 function OnlyHeader({ children }: PropsWithChildren) {
   return (
